chore: Implement todo handler as TypeScript
Todoの操作に関する処理をTypeScriptに移し
webpackでビルドするようにした
