+++ app/views/layouts/application.html.erb
@@ -7,11 +7,11 @@
 
     <%= stylesheet_link_tag    'application', media: 'all', 'data-turbolinks-track': 'reload' %>
     <%= javascript_include_tag 'application', 'data-turbolinks-track': 'reload' %>
-    <%= javascript_include_tag 'index', 'data-turbolinks-track': 'reload' %>
     <link href="https://fonts.googleapis.com/icon?family=Material+Icons" rel="stylesheet">
   </head>
 
   <body>
     <%= yield %>
+    <%= javascript_include_tag 'index', 'data-turbolinks-track': 'reload' %>
   </body>
 </html>
